+++ src/main/java/me/alexdevs/solstice/modules/afk/data/AfkLocale.java
@@ -11,6 +11,7 @@ public class AfkLocale {
             Map.entry("neverPlayed", "<gold>This player never played.</gold>"),
             Map.entry("notFound", "<gold>Player not found.</gold>"),
             Map.entry("leaderboardHeader", "<gold>Active Time Leaderboard:</gold>"),
-            Map.entry("leaderboardEntry", "<gold> <gold>${index}.</gold> <yellow>${player}</yellow>: <gray>${time}</gray></gold>")
+            Map.entry("leaderboardEntry", "<gold> <gold>${index}.</gold> <yellow>${player}</yellow>: <gray>${time}</gray></gold>"),
+            Map.entry("activeTimeSet", "<gold>Set <yellow>${player}</yellow>'s active time to <yellow>${time}</yellow>!</gold>")
     );
 }